backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic[email]==2.6.1
sqlalchemy==2.0.27
alembic==1.13.1
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
python-dotenv==1.0.1
slowapi==0.1.9

# ── Testing (dev only) ────────────────────────────────────────────────────────
pytest==8.1.1
pytest-asyncio==0.23.6
anyio==4.3.0
